+++ packages/bento-web/src/profile/ProfilePage/components/QuestionMoreModal.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 
 import { Modal } from '../../components/Modal';
-import { Portal } from '../../components/Portal';
+import { Portal } from '@/components/Portal';
 
 type Props = {
   isVisible?: boolean;
